BrowseLanguage code cleaned
