+++ src/test/java/pages/browse_languages/languages/AdaLanguagePage.java
@@ -1,9 +1,8 @@
 package pages.browse_languages.languages;
 
 import org.openqa.selenium.WebDriver;
-import pages.browse_languages.letters.LetterPage;
 
-public class AdaLanguagePage extends LetterPage {
+public class AdaLanguagePage extends LanguagePage {
 
     public AdaLanguagePage(WebDriver driver) {
         super(driver);
