+++ src/test/java/pages/browse_languages/languages/JQueryLanguagePage.java
@@ -1,7 +1,6 @@
 package pages.browse_languages.languages;
 
 import org.openqa.selenium.WebDriver;
-import pages.browse_languages.BrowseLanguagesSubmenuPage;
 
 public class JQueryLanguagePage extends LanguagePage {
