+++ src/test/java/pages/browse_languages/languages/JavaLanguagePage.java
@@ -1,9 +1,6 @@
 package pages.browse_languages.languages;
 
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import pages.base_abstract.BasePage;
 
 public class JavaLanguagePage extends LanguagePage {
