+++ src/test/java/pages/browse_languages/languages/JavaScriptLanguagePage.java
@@ -3,11 +3,10 @@ package pages.browse_languages.languages;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import pages.base_abstract.BasePage;
 
 public class JavaScriptLanguagePage extends LanguagePage {
 
-    @FindBy(xpath = "//*[@id='main']//a[contains(text(), 'http://en.wikipedia.org/wiki/Javascript')]")
+    @FindBy(xpath = "//div[@id='main']//a[contains(text(), 'http://en.wikipedia.org/wiki/Javascript')]")
     private WebElement JavaScriptDeepLink;
 
     public JavaScriptLanguagePage(WebDriver driver) {
