+++ src/test/java/pages/browse_languages/languages/ShakespeareLanguagePage.java
@@ -1,7 +1,6 @@
 package pages.browse_languages.languages;
 
 import org.openqa.selenium.WebDriver;
-import pages.top_lists.TopListsSubmenuPage;
 
 public class ShakespeareLanguagePage extends LanguagePage {
