+++ src/test/java/pages/browse_languages/letters/ABCPage.java
@@ -3,12 +3,8 @@ package pages.browse_languages.letters;
 import org.openqa.selenium.WebDriver;
 
 public class ABCPage extends LetterPage {
+
     public ABCPage(WebDriver driver) {
         super(driver);
-    }
-
-    public String getABCPageURL() {
-
-        return getURL();
     }
 }
