+++ src/test/java/tests/Ntest.java
@@ -18,7 +18,7 @@ public class Ntest extends BaseTest {
 
         String oldURL = openBaseURL()
                 .clickBrowseLanguagesMenu()
-                .getABCPageURL();
+                .getURL();
 
         new ABCPage(getDriver()).clickNSubmenu();
